[feat] #40 Entity 관계 구성
