+++ src/main/java/DGU_AI_LAB/admin_be/domain/nodes/NodeStatus.java
@@ -1,4 +1,4 @@
-package DGU_AI_LAB.admin_be.domain.users.entity;
+package DGU_AI_LAB.admin_be.domain.nodes;
 
 public enum NodeStatus {
     ACTIVE, INACTIVE, ERROR
